// src/app/model/cita.ts
/**
 * ARCA
 * Aqui descripcion de la api
 *
 * OpenAPI spec version: 0.0.1
 * Contact: no hay
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Veterinario } from './veterinario';

export interface Cita { 
    estado: boolean;
    fechaCita: Date;
    id?: number;
    motivo: string;
    nombreCliente: string;
    veterinario?: Veterinario;
}